RESPALDO_EC2_MAY15/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
reportlab==4.4.0
Pillow==11.2.1
